ipamd: Move main.go to cmd/ipamd.
A common idiom used by many projects in the Go community is to place
all commands in a directory called cmd. This change introduces this
repository to this approach.

There are many projects that use this (related: kubernetes, kops), but
the general thinking behind it is discussed in this document:

https://dave.cheney.net/2014/12/01/five-suggestions-for-setting-up-a-go-project