Add logs and metrics for cooling IP count
When a node is near full capacity in terms of IP addresses used, then logs can
become confusing for understanding why ipamD was unable to assign an IP to
a new pod. For example, the added test case would previously report the
following in logs:

```
[Debug] UnAssignIPv4Address: IP address pool stats: total:1, assigned 1, pod(Name: pod-1, Namespace: ns-1, Container container-1)
[Info] UnAssignIPv4Address: pod (Name: pod-1, NameSpace ns-1 Container container-1)'s ipAddr 1.1.1.1, DeviceNumber1
[Debug] AssignIPv4Address: IP address pool stats: total:1, assigned 0
[Info] DataStore has no available IP addresses
```

The 3rd line indicates that there's an IP available, but the 4th one concludes
that no IPs are available. It is correct that no IPs are available, because all
unused IPs are in ["cooling mode"][1]. The problem is that that is not obvious
from the logs. With these changes, the logs for the test are changed to the
following:

```
[Debug] UnAssignIPv4Address: pod(Name: pod-1, Namespace: ns-1, Container container-1), IP address pool stats: total: 1, assigned: 1, cooling: 0
[Info] UnAssignIPv4Address: pod (Name: pod-1, NameSpace ns-1 Container container-1)'s ipAddr 1.1.1.1, DeviceNumber 1
[Debug] AssignIPv4Address, IP address pool stats: total: 1, assigned: 0, cooling: 1
[Info] DataStore has no available IP addresses
```

This makes it clear why IP allocation failed.

Commit 3ccfef90 ("Add 30 sec cooling for IP address reassignment", 2018-04-13)
added a cooling period for IPs, but the behavior isn't currently covered by
tests, so a test was added. Ideally the fact that the IP is actually usable
after the cooling period would also be tested, but this would require either
refactoring the code to inject a mock `time.Now()` or including a dependency to
the project to monkey patch the same (or sleeping in test :S). Not doing these
changes right now to keep the patches minimal.

[1]: https://github.com/aws/amazon-vpc-cni-k8s/blob/3735b4a5c38524504cf11429bd731fa833256d2f/docs/cni-proposal.md#pod-ip-address-cooling-period