Remove unExpectedMaxENIAttached error case
Currently only a metric is emitted.  This should be removed because
I don't think we care when this happens and it's clutter.

The case is true when the discovered c.maxENI (which happens when
an ENI fails to be allocated due to maxENIs already present on the
instance) is less than maxENIs (either the env var or the maximum
allowed by the instance).