Allow additional CIDRs to be excluded from SNAT (#520)
* Allow additional CIDRs to be excluded from SNAT

Introduces a mechanism to provide an exclusion list of CIDRs that should
not be SNATed.

Adds an environment variable `AWS_VPC_K8S_CNI_EXCLUDE_SNAT_CIDRS` to
provide a comma separated list of ipv4 CIDRs to exclude from SNAT.
The value of this environment variable will only be use when
`AWS_VPC_K8S_CNI_EXTERNALSNAT` is false.

The SNAT exclusion CIDRs will be used to:
- Generate `iptable` rules to prevent SNATing for packets directed to
the excluded CIDRs
- Generate `ip rule` entries to route packets directed to the excluded
CIDRs through the pod's `eni`. These configuration is done both at the
`cni` plugin level via the `rpc_handler`, and the runtime `network` api.

Given that the list of excluded CIDRs may vary by configuration it was
necessary to include a mechanism to clean up stale SNAT rules. If a
CIDR is removed from the exclusion list, the corresponding `iptable`
rule will be removed as well, and the chain will be adjusted.

Connects https://github.com/aws/amazon-vpc-cni-k8s/issues/469

Co-authored-by: @rewiko
Co-authored-by: @yorg1st

* Add formatting targets to Makefile

`format` applies formatting to the project's go files.
`check-format` checks that no files require formatting; if they do it
will fail the command.

It adds the `check-format` target to the travis build so that CI fails
if files are not properly formatted.

* Fix typo in SNAT chain rules' comment