Scope watch on "pods" to only pods associated with the local node
Prior to this patch the aws-cni plugin would create a watch for *all*
pods in the cluster to then client-side filter them to only the ones
assigned to the node. This is not only wasteful of resources (as we're
throwing away the majority of data once we get it) this causes some
significant performance/capacity issues on sizeable clusters. In some
cases we've seen the initial `watch` call return ~500MB! This is a
significant savings of space and resources (both client and server side)