do pre-checks before creating EKS test cluster
Ensure that there is an ECR repo before we use aws-k8s-tester to actually
create the test cluster. In addition, trap errors in the main
run-integration-tests.sh script and deprovision cluster if an error
occurred after creating the cluster.

Updates the required aws-k8s-tester release to 0.5.2 to handle
aws/aws-k8s-tester#66

Issue #765