make datastore.GetENINeedsIPs deterministic
The datastore used in ipamd uses a map[string]*IPPool to store ENI
information. The `datastore.GetENINeedsIP()` method was iterating over
the map using `range`, which is not deterministic since a map's elements
have no guaranteed order.

This was problematic for a unit test since a mock was expecting the
primary ENI to be returned from the datastore but every so often, the
secondary ENI was being returned due to the non-deterministic map
iteration. This patch fixes this by simply sorting the map's keys and
iterating over the sorted keys, accessing the ENI information by key
instead by iteration.

Fixes Issue #814