ensure non-nil Attachement in getENIAttachmentID
Practice good code safety in the `EC2MetadataCache.getENIAttachmentID()`
method by not assuming that either the `DescribeNetworkInterfacesOutput`
struct's `NetworkInterfaces` field is not empty and that the first
`NetworkInterface` struct that collection has a non-nil `Attachment`
field.

Fixes Issue #914 however note that with #909, the source code changed
dramatically and this patch will need to be written differently for
v1.5.x branches.