no longer refer to potential CLA requirement
Our legal team has asked that the wording included in the
CONTRIBUTING.md document referring to potentially requiring a CLA to be
signed for certain contributions be removed.

This phrase was added out of an abundance of caution originally and has
been determined to be unnecessary and inhibiting external contribution.