aws-node does not perform ip unassignment properly
Sometimes aws-node pod from aws-node daemonset does not perform unassignment of non-existing docker containers.
This commit marks IPAMKey data structures where ContainerID is empty as unassigned.

Example of ipamd output:
"192.168.9.204": {
  "Address": "192.168.9.204",
  "IPAMKey": {
    "containerID": "d74e85090b40c25325b82289ad335287ab9467784ea4014052e49f3032c4e9e8",
    "ifName": "unknown",
    "networkName": "_migrated-from-cri"
  },
  "UnassignedTime": "0001-01-01T00:00:00Z"
},
"192.168.9.25": {
  "Address": "192.168.9.25",
  "IPAMKey": {
    "containerID": "",
    "ifName": "",
    "networkName": ""
  },
  "UnassignedTime": "0001-01-01T00:00:00Z"
},
"192.168.9.46": {
  "Address": "192.168.9.46",
  "IPAMKey": {
    "containerID": "",
    "ifName": "",
    "networkName": ""
  },
  "UnassignedTime": "0001-01-01T00:00:00Z"
},

Eventually we are facing situation when we have free IP addresses but because ipamd thinks it is still assigned we are not able to add new containers.
After we re-deploy aws-node daemonset everything is back to normal again for couple of days.