chore: Add memory request for avoiding EC2 OOM
When no memory request is specified for any pod running on the EC2's, the EC2 instance ends up going in to OOM and becoming non-responsive, leaving the kubelet to crash and leave the node in a not-ready state. Adding a minimum request for memory will allow VPC CNI to be always available and potentially trigger evictions for other application pods without CPU/Memory.

The 200Mi has been set based on my observation across an average of 100 nodes used over two weeks.