Prioritize lower DeviceNumber in ipv4 assignment
Currently (*DataStore).AssignPodIPv4Address() iterates through the
eniPool map when searching for an IP address to assign to the pod.
Because map iteration order is random, this means that new assignments
will tend to be distributed evenly across all ENIs with sufficient
available IPs.

When WARM_ENI_TARGET is set (and not {WARM,MINIMUM}_IP_TARGET), IPs are
only returned to the subnet when the entire ENI can be removed, which
can only happen when it's completely unused.

For our workloads, even distribution of new assignments makes it very
unlikely for any ENI to become unused, once added. This in turn means
that ipamd almost never returns IPs to the subnet, causing us to have
many more IPs allocated than can ever be used in practice.